Refactor Term/Grid separation, Add regex scrollback buffer search
This PR aims to clear up the separation between Term and Grid to make
way for the search PR.

The `cursor` and `cursor_save` have been moved to the grid, since
they're always bound to their specific grid and this makes updating
easier.

Since the selection is indepedent of the active grid, it has been moved
to the `Term`.